Move Meet_ILT classes to prototyping
Rename/move several Java classes from org.firstinspires.ftc.teamcode.Meet_ILT (and its Teleop subpackage) into org.firstinspires.ftc.teamcode.prototyping. Updated package declarations in Configurables, ILT_Teleop, InputHandler, Motors and StateMachine, and removed now-unused imports/references (e.g. SparkFunOTOS, some Teleop-specific imports) during the move to clean up code and resolve packaging.
