+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/prototyping/Configurables.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Meet_ILT.Teleop;
+package org.firstinspires.ftc.teamcode.prototyping;
 
 import com.bylazar.configurables.annotations.Configurable;
